refactor: Move `utils.ts` -> `./lib`
